# back-end/server/docs/swagger.yaml
basePath: /
definitions:
...
  models.CarInsuranceForm:
    properties:
      car_form_id:
        type: integer
      chassis_number:
        type: string
      created_at:
        type: string
      engine_number:
        type: string
      form_id:
        type: integer
      identity_number:
        type: string
      insurance_amount:
        type: number
      insurance_duration:
        type: integer
      insurance_fee:
        type: number
      insurance_start:
        type: string
      license_plate:
        type: string
      license_plate_status:
        type: string
      load_capacity:
        type: number
      owner_name:
        type: string
      participant_count:
        type: integer
      registration_address:
        type: string
      seat_count:
        type: integer
      updated_at:
        type: string
      usage_purpose:
        type: string
      user_type:
        type: string
      vehicle_type:
        type: string
    type: object
...
  models.CustomerRegistration:
    properties:
      address:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      customer_type:
        type: string
      email:
        type: string
      full_name:
        type: string
      identity_number:
        type: string
      invoice_request:
        type: boolean
      notes:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
...
  models.PersonalInsuranceForm:
    properties:
      birth_date:
        type: string
      cmnd_img:
        type: string
      created_at:
        type: string
      dental_extension:
        type: boolean
      form_id:
        description: Cho phép NULL
        type: integer
      full_name:
        type: string
      gender:
        type: string
      identity_number:
        type: string
      insurance_duration:
        type: integer
      insurance_fee:
        type: number
      insurance_program:
        type: string
      insurance_start:
        type: string
      maternity_extension:
        type: boolean
      personal_form_id:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: API cho hệ thống bảo hiểm
  title: BIC Insurance API
  version: "1.0"
paths:
...
  /api/insurance_cancer/create_customer_registration:
    post:
      consumes:
      - application/json
      parameters:
      - description: Customer info
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.CustomerRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Đăng ký khách hàng
      tags:
      - Customer
...
  /api/insurance_car_owner/create_car_insurance_form:
    post:
      consumes:
      - application/json
      parameters:
      - description: Car insurance form
        in: body
        name: carInsuranceForm
        required: true
        schema:
          $ref: '#/definitions/models.CarInsuranceForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Lưu thông tin bảo hiểm xe ô tô
      tags:
      - CarOwner
  /api/insurance_car_owner/create_customer_registration:
    post:
      consumes:
      - application/json
      parameters:
      - description: Customer info
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.CustomerRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Đăng ký khách hàng
      tags:
      - Customer
...
  /api/insurance_home/create_customer_registration:
    post:
      consumes:
      - application/json
      parameters:
      - description: Customer info
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.CustomerRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Đăng ký khách hàng
      tags:
      - Customer
...
  /api/insurance_motorbike_owner/create_customer_registration:
    post:
      consumes:
      - application/json
      parameters:
      - description: Customer info
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.CustomerRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Đăng ký khách hàng
      tags:
      - Customer
...
  /api/insurance_personal/create_customer_registration:
    post:
      consumes:
      - application/json
      parameters:
      - description: Customer info
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.CustomerRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Đăng ký khách hàng
      tags:
      - Customer
...
  /api/insurance_personal/create_personal_insurance_form:
    post:
      consumes:
      - application/json
      parameters:
      - description: Personal insurance form
        in: body
        name: personalInsuranceForm
        required: true
        schema:
          $ref: '#/definitions/models.PersonalInsuranceForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Lưu bảo hiểm sức khỏe cá nhân
      tags:
      - Personal
  /api/insurance_travel/create_customer_registration:
    post:
      consumes:
      - application/json
      parameters:
      - description: Customer info
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.CustomerRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Đăng ký khách hàng
      tags:
      - Customer
...
swagger: "2.0"